#173 Keep labelled_map as it provides num active steps
