+++ baal/active/dataset/__init__.py
@@ -1,0 +1,4 @@
+from .numpy import ActiveNumpyArray
+from .pytorch_dataset import ActiveLearningDataset
+
+# Do not include HF Dataset here as it requires more deps.
